Make BinaryCollision a templated class with the functor type as parameter
