+++ Source/Particles/Collision/CMakeLists.txt
@@ -2,6 +2,5 @@ target_sources(WarpX
   PRIVATE
     CollisionHandler.cpp
     CollisionBase.cpp
-    BinaryCollision.cpp
 )
