+++ Source/Particles/Collision/CollisionHandler.cpp
@@ -7,6 +7,7 @@
 #include "CollisionHandler.H"
 
 #include "BinaryCollision.H"
+#include "PairWiseCoulombCollisionFunc.H"
 
 #include <AMReX_ParmParse.H>
 
@@ -33,7 +34,8 @@ CollisionHandler::CollisionHandler()
         collision_types[i] = type;
 
         if (type == "pairwisecoulomb") {
-            allcollisions[i] = std::make_unique<BinaryCollision>(collision_names[i]);
+            allcollisions[i] =
+               std::make_unique<BinaryCollision<PairWiseCoulombCollisionFunc>>(collision_names[i]);
         }
 
     }
