+++ Source/Particles/Collision/Make.package
@@ -1,5 +1,4 @@
 CEXE_sources += CollisionHandler.cpp
 CEXE_sources += CollisionBase.cpp
-CEXE_sources += BinaryCollision.cpp
 
 VPATH_LOCATIONS   += $(WARPX_HOME)/Source/Particles/Collision
